+++ changeLog.md
@@ -1,3 +1,7 @@
 ```
 - first approach to insertTask function :scream:
 ```
+
+```
+- newsstand restrict duplicates on business name only (not business and piva anymore)
+```